Add the ability to sample multiple signals
Samplers can now provide a list of available signals, and the GUI can set
the active signals for each sampler.
The available signals are displayed in the GUI and can be selected to be active.

Other minor changes.

Currently a bit unstable, needs lots of improvements, and must only be
considered a quick sketch of the architecture.