fix(openocd/telnet_interface): more robust response reading
Do not just read the following line and check that it suits the wanted
response, but keep reading lines until one that matches the predicate is
found (or a timeout happens).

Also collect all of the discarded lines and
return them in an `UnexpectedResponse` Telnet error so that their content
is propagated and inspected for error handling purposes.